% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correspond_absmaps.R
\name{correspond_absmaps}
\alias{correspond_absmaps}
\title{Read in ASGC/ASGS population-weighted correspondence tables from the ABS.}
\arguments{
\item{data}{The original data.frame or tibble containing  that you want 'corresponded'.}

\item{fromArea}{The original area to correspond from ("sa1", "sa2", etc; see documentation for the full list.)}

\item{fromYear}{The year to correspond from (2011, 2016, etc; see documentation for full list.)}

\item{toArea}{The original area to correspond to ("sa1", "sa2", etc; see documentation for the full list.)}

\item{toYear}{The year to correspond to (2011, 2016, etc; see documentation for full list.)}

\item{nVar}{The variable you want distributed according to the correspondence ratio.
For example, \code{count} or \code{n} that represent the number of people in a given area.}
}
\value{
A \code{tibble} object.
}
\description{
Read in ASGC/ASGS population-weighted correspondence tables from the ABS.
}
\examples{
\dontrun{

}
}
