% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_abs_shapefile.R
\name{download_abs_shapefile}
\alias{download_abs_shapefile}
\alias{download_abs_shapefiles}
\title{Download ABS ASGS and/or ASGC shapefiles to a specified path}
\usage{
download_abs_shapefiles()
}
\description{
Currently, this function will only work inside
\code{download_absmaps} and \code{load_absmaps}.
}
