% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_correspondence_absmaps.R
\name{get_correspondence_absmaps}
\alias{get_correspondence_absmaps}
\title{Read in ASGC/ASGS population-weighted correspondence tables from the ABS.}
\usage{
get_correspondence_absmaps(from, fromyear, to, toyear,
  type = "onetomany")
}
\arguments{
\item{from}{The area you want to correspond FROM (ie the areas your data are currently in). For example: "sa1", "sa2, "sa3", "sa4".}

\item{fromyear}{The year you want to correspond FROM. For example: 2011, 2016.}

\item{to}{The area you want to correspond TO (ie the areas you want your data to be in).}

\item{toyear}{The year you want to correspond TO.}

\item{type}{Either 'onetomany' (the default) or 'manytomany'. 'onetomany' filters the observations to keep the best single match for each FROM area.}
}
\value{
A \code{tibble} object.
}
\description{
Read in ASGC/ASGS population-weighted correspondence tables from the ABS.
}
\examples{
\dontrun{
sa4_corr <- get_correspondence_absmaps("sa4", 2011, "sa4", 2011)
}
}
