% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_and_compress_shapefile.R
\name{read_and_compress_shapefile}
\alias{read_and_compress_shapefile}
\title{Read and compress ABS ASGS and/or ASGC shapefile
data, compress and save as an \code{sf} object.}
\usage{
read_and_compress_shapefile()
}
\description{
Currently, this function will only work inside
\code{download_absmaps} and \code{load_absmaps}.
}
