% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correspond_absmaps.R
\name{correspond_absmaps}
\alias{correspond_absmaps}
\title{Read in ASGC/ASGS population-weighted correspondence tables from the ABS.}
\usage{
correspond_absmaps(data, fromArea, fromYear, toArea, toYear, nVar)
}
\arguments{
\item{data}{The original data.frame or tibble containing  that you want 'corresponded'.}

\item{nVar}{The variable you want distributed according to the correspondence ratio.
For example, \code{count} or \code{n} that represent the number of people in a given area.}
}
\value{
A \code{tibble} object.
}
\description{
Read in ASGC/ASGS population-weighted correspondence tables from the ABS.
}
\examples{
\dontrun{

}
}
