% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_absmaps.R
\name{download_absmaps}
\alias{download_absmaps}
\title{Download ABS ASGS and/or ASGC shapefile data, compress and save
as an \code{sf} object.}
\arguments{
\item{area}{The area you want to download/convert/compress. One of
"sa1", "sa2, "sa3", "sa4", "gcc", "state".}

\item{year}{The year of the ASGS data. Defaults to 2016, but 2011 will
be useful for older data.}

\item{saveDirectory}{The path to which your map data is saved. Default is
the current working directory.}

\item{mapCompression}{The compression level of your map data. Default
is 0.1 -- 10 per cent of original detail -- which makes clear, detailed maps. Higher
compression leads to greater map file size with, in most cases, little visual benefit.
Map compression can be set to "off".}

\item{removeSourceFiles}{Remove the original ABS shapefile data after
compression. Defaults to TRUE.}
}
\description{
Download ABS ASGS and/or ASGC shapefile data, compress and save
as an \code{sf} object.
}
\examples{
\dontrun{
download_absmaps("sa3", 2016)
}
}
